TASK: refactor e2e testing setup (#2403)
