+++ Tests/IntegrationTests/Fixtures/1Dimension/SitePackage/Configuration/NodeTypes.Content.Headline.yaml
@@ -1,0 +1,24 @@
+'Neos.TestSite:Content.Headline':
+  superTypes:
+    'Neos.Neos:Content': true
+  ui:
+    label: Headline
+    icon: icon-header
+    position: 200
+    inlineEditable: true
+  properties:
+    title:
+      type: string
+      defaultValue: '<h1>Enter headline here</h1>'
+      ui:
+        inlineEditable: true
+        inline:
+          editorOptions:
+            placeholder: '<h2>Enter headline here...</h2>'
+            formatting:
+              h1: true
+              h2: true
+              h3: true
+              h4: true
+              h5: true
+              a: true
