+++ Tests/IntegrationTests/Fixtures/1Dimension/SitePackage/Configuration/NodeTypes.Content.Image.yaml
@@ -1,0 +1,30 @@
+'Neos.TestSite:Content.Image':
+  superTypes:
+    'Neos.Neos:Content': true
+  ui:
+    label: Image
+    icon: icon-picture
+    position: 300
+    inspector:
+      groups:
+        image:
+          label: 'Image'
+          position: 5
+          icon: 'icon-image'
+  properties:
+    image:
+      type: Neos\Media\Domain\Model\ImageInterface
+      ui:
+        label: 'Image'
+        reloadIfChanged: true
+        inspector:
+          group: 'image'
+          position: 50
+    alternativeText:
+      type: string
+      ui:
+        label: 'Alternative Text'
+        reloadIfChanged: false
+        inspector:
+          group: 'image'
+          position: 100
