+++ Tests/IntegrationTests/Fixtures/1Dimension/SitePackage/Configuration/NodeTypes.Document.Page.yaml
@@ -1,0 +1,26 @@
+'Neos.TestSite:Document.Page':
+  superTypes:
+    'Neos.Neos:Document': true
+  ui:
+    label: Page
+    icon: icon-file-o
+    position: 100
+  childNodes:
+    main:
+      type: 'Neos.Neos:ContentCollection'
+  properties:
+    image:
+      type: Neos\Media\Domain\Model\ImageInterface
+      ui:
+        label: 'Image'
+        reloadIfChanged: true
+        inspector:
+          group: 'document'
+          editorOptions:
+            features:
+              crop: true
+            crop:
+              aspectRatio:
+                locked:
+                  width: 1
+                  height: 1
