+++ Tests/IntegrationTests/Fixtures/1Dimension/SitePackage/Resources/Private/Fusion/Content/Headline/Headline.fusion
@@ -1,0 +1,9 @@
+prototype(Neos.TestSite:Content.Headline) < prototype(Neos.Neos:ContentComponent) {
+    title = Neos.Neos:Editable {
+        property = 'title'
+    }
+
+    renderer = afx`
+        <div class="test-headline">{props.title}</div>
+    `
+}
