+++ Tests/IntegrationTests/Fixtures/1Dimension/SitePackage/Resources/Private/Fusion/Content/Text/Text.fusion
@@ -1,0 +1,9 @@
+prototype(Neos.TestSite:Content.Text) < prototype(Neos.Neos:ContentComponent) {
+    text = Neos.Neos:Editable {
+        property = 'text'
+    }
+
+    renderer = afx`
+        <div class="test-text">{props.text}</div>
+    `
+}
