+++ Tests/IntegrationTests/Fixtures/1Dimension/SitePackage/Resources/Private/Fusion/Root.fusion
@@ -1,0 +1,2 @@
+include: **/*.fusion
+page = Neos.TestSite:Document.Page
