+++ Tests/IntegrationTests/Fixtures/1Dimension/SitePackage/composer.json
@@ -1,0 +1,15 @@
+{
+    "name": "neos/test-site",
+    "description": "A Neos website",
+    "type": "neos-site",
+    "require": {
+        "neos/neos": "*",
+        "neos/neos-ui": "*",
+        "neos/neos-ui-compiled": "*"
+    },
+    "extra": {
+        "neos": {
+            "package-key": "Neos.TestSite"
+        }
+    }
+}
