+++ Tests/IntegrationTests/TestDistribution/.gitignore
@@ -1,0 +1,9 @@
+Packages
+Web
+Configuration/*
+!Configuration/Settings.yaml
+Data
+bin
+Readme.rst
+Upgrading.rst
+flow*
