+++ Tests/IntegrationTests/TestDistribution/Configuration/Settings.yaml
@@ -8,6 +8,11 @@ Neos:
         dbname: neos
         user: root
         host: 127.0.0.1
+        password: not_a_real_password
+    reflection:
+      ignoredTags:
+        template: true
+        psalm: true
   Neos:
     Ui:
       frontendDevelopmentMode: true
