+++ packages/react-ui-components/src/Tree/reactDndScrollzoneFork.js
@@ -2,6 +2,8 @@
 // https://github.com/azuqua/react-dnd-scrollzone/pull/38#issuecomment-434385410
 
 import React, {Component, createContext} from 'react';
+// WARNING: this needs to be changed on newer react-dnd versions to this:
+// import {DragDropContextConsumer} from 'react-dnd';
 import {Consumer as DragDropContextConsumer} from 'react-dnd/lib/DragDropContext';
 import PropTypes from 'prop-types';
 import {findDOMNode} from 'react-dom';
